Scripts/requirements.txt
openai>=1.23.0
python-dotenv>=1.0.1
requests>=2.31.0
